requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==3.2.0
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==24.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
discord-py==2.4.0
frozenlist==1.5.0
google-auth==2.43.0
google-genai==1.50.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.1.0
numpy==1.26.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.4
pydantic-core==2.41.5
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.18.3
